pyproject.toml

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "authscope"
version = "0.1.0"
description = "Authorization invariants tester for Web/API (BOLA/BFLA)"
authors = [{name = "Ahmad Kabiri"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.6",
    "PyYAML>=6.0",
    "typer>=0.12",
    "rich>=13.7",
]

[project.urls]
Homepage = "https://github.com/ahmad-kabiri/authscope"
Issues = "https://github.com/ahmad-kabiri/authscope/issues"

[project.scripts]
authscope = "authscope.cli:main"
